Able to calculate cut sets of fault tree and fixed logic with FTA
Created methods to calculate the cut set of the fault tree from just the data generated by the fault tree. Fixed a logical mistake with the fault tree gates, since failures use reverse logic. Therefore AND is OR in faults and OR is AND in faults. Also added new test data sets.